Update sketch.js
Added 'function changeDiceValue()' with random and parse functions to generate whole numbers for roll of dice. parseInt lops off the decimal place eg 6.87 becomes 6.
Added  'if(DiceValue == 1){  DiceValueOne(40, 20, 20);  } to show Dice face one when 1 is randomly generated from the 'Click Here' button. The '==' and '!=' are 2 of the Equality operators within JavaScript Expressions and operators. Notes: Be careful with curly brackets for readability. Be careful of correct Scope for sections of code.